@@ -69,7 +69,7 @@
     "path-to-regexp": "^6.2.1",
     "source-map-support": "^0.5.21",
     "tslib": "^2.5.0",
-    "typescript": "^4.9.5",
+    "typescript": "^5.0.0-beta",
     "ws": "^8.12.1"
   }
 }
