@@ -1,7 +1,5 @@
 export abstract class Gate {
   public allows(action: string, data: unknown): boolean {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    return this[action]?.(data) ?? false;
+    return (this as unknown as Record<string, Function>)[action]?.(data) ?? false;
   }
 }
