@@ -3,9 +3,9 @@ import { RouteUrl } from '../../router/types/route-url.type.js';
 import { ClassDecorator } from '../../utils/types/class-decorator.type.js';
 
 export function Controller(baseUrl?: RouteUrl): ClassDecorator {
-  return (target) => {
-    Reflect.defineMetadata('baseUrl', baseUrl, target);
+  return (originalClass) => {
+    Reflect.defineMetadata('baseUrl', baseUrl, originalClass);
 
-    return target;
+    return originalClass;
   };
 }
