@@ -3,7 +3,9 @@ import { Constructor } from '../../utils/interfaces/constructor.interface.js';
 import { MethodDecorator } from '../../utils/types/method-decorator.type.js';
 
 export function Middleware(middleware: Constructor): MethodDecorator {
-  return (target) => {
-    Reflect.defineMetadata('middleware', middleware, target);
+  return (originalClass) => {
+    Reflect.defineMetadata('middleware', middleware, originalClass);
+
+    return originalClass;
   };
 }
