@@ -3,7 +3,9 @@ import { MethodDecorator } from '../../utils/types/method-decorator.type.js';
 import { StatusCode } from '../enums/status-code.enum.js';
 
 export function Status(status: StatusCode = StatusCode.Ok): MethodDecorator {
-  return (target) => {
-    Reflect.defineMetadata('statusCode', status, target);
+  return (originalClass) => {
+    Reflect.defineMetadata('statusCode', status, originalClass);
+
+    return originalClass;
   };
 }
