@@ -2,7 +2,9 @@ import { Reflection as Reflect } from '@abraham/reflection';
 import { MethodDecorator } from '../../utils/types/method-decorator.type.js';
 
 export function Name(name: string): MethodDecorator {
-  return (target) => {
-    Reflect.defineMetadata('name', name, target);
+  return (originalClass) => {
+    Reflect.defineMetadata('name', name, originalClass);
+
+    return originalClass;
   };
 }
