@@ -3,7 +3,9 @@ import { Integer } from '../../utils/types/integer.type.js';
 import { MethodDecorator } from '../../utils/types/method-decorator.type.js';
 
 export function Throttle(maxRequestsPerMinute: Integer): MethodDecorator {
-  return (target) => {
-    Reflect.defineMetadata('maxRequestsPerMinute', maxRequestsPerMinute, target);
+  return (originalClass) => {
+    Reflect.defineMetadata('maxRequestsPerMinute', maxRequestsPerMinute, originalClass);
+
+    return originalClass;
   };
 }
