@@ -4,8 +4,8 @@ import { ClassDecorator } from '../../utils/types/class-decorator.type.js';
 export function Channel(name: string, serverName = 'main'): ClassDecorator {
   const pattern = pathToRegexp(name);
 
-  return (target) => {
-    return class extends target {
+  return (originalClass) => {
+    return class extends originalClass {
       public readonly namePattern = pattern;
 
       public readonly serverName = serverName;
